+++ test/every.test.js
@@ -4,7 +4,7 @@ const spacetime = require('./lib')
 
 test('every-unit', t => {
   let start = spacetime('April 6th 2019', 'Europe/Paris')
-  let end = spacetime('April 20th 2019', 'Europe/Paris')
+  let end = spacetime('April 20th 2019', 'Europe/Paris').add(1, 'hour')
 
   let days = start.every('day', end)
   t.equal(days.length, 14, '14 days')
@@ -23,6 +23,7 @@ test('monday-sunday', t => {
   let days = ['monday', 'tuesday', 'wednesday', 'thursday', 'friday', 'saturday', 'sunday']
   let start = spacetime('April 8th 2019').startOf('week')
   let end = start.endOf('week')
+  start = start.minus(1, 'hour')
   let eachDay = start.every('day', end).map(d => d.dayName())
   t.deepEqual(eachDay, days, 'got mon-sunday')
   t.end()