+++ changelog.md
@@ -46,6 +46,7 @@
 - avoid setting dates/months/hours past maximums and minimums
 - `.hour(24)` now changes date to tomorrow
 - faster diffs for large (decade-long) ranges
+- **[big]** fix awkward diff/since issues ('dec 25->jan 5' should be 0 years)
 
 ### v4.0.0